+++ R/zzz.R
@@ -1,0 +1,19 @@
+.onLoad <- function(libname, pkgname) {
+  op <- options()
+  op.TROLL <- list(
+    TROLL.path = "~/Documents/ECOFOG/Results",
+    TROLL.src = system.file("tools", "main.cpp",  
+                            package = 'TROLL'),
+    TROLL.app = "TROLL.out",
+    TROLL.init = "input.txt",
+    TROLL.forest = "forest.txt",
+    TROLL.species = system.file("extdata", "species.txt",  
+                                package = 'TROLL'),
+    TROLL.missing = "missing.txt",
+    TROLL.output = "OUTPUT"
+  )
+  toset <- !(names(op.TROLL) %in% names(op))
+  if(any(toset)) options(op.TROLL[toset])
+  
+  invisible()
+}